Hot-Fix_LastYear main.js
Hotfix for Senec machines not older than one year, respectively installed in the current year.
The adapter cannot retrieve values for LAST_YEAR and receives NULL.
As a hotfix, I commented out LAST_YEAR in the date MAP.
This will stop the errors in the protocol and allow the other API values to continue working as intended.

This is just a temporary emergency solution to get the adapter continue working. For the future, a check must be impelemented that automatically checks and corrects this error for machines < 1 year old.